# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "PolyGo API",
    "description": "High-performance Polymarket API proxy with caching and WebSocket support",
    "version": "1.0",
    "contact": {
      "name": "API Support",
      "email": "support@polygo.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "POLY-API-KEY"
    }
  },
  "paths": {
...
    "/api/v1/markets": {
      "get": {
        "tags": ["Markets"],
        "summary": "List all markets",
        "description": "Get a list of markets with optional filtering",
        "produces": ["application/json"],
        "parameters": [
          {"name": "limit", "in": "query", "type": "integer", "default": 100},
          {"name": "cursor", "in": "query", "type": "string"},
          {"name": "active", "in": "query", "type": "boolean"},
          {"name": "closed", "in": "query", "type": "boolean"}
        ],
        "responses": {
          "200": {"description": "List of markets"}
        }
      }
    },
    "/api/v1/markets/{id}": {
      "get": {
        "tags": ["Markets"],
        "summary": "Get market by ID",
        "produces": ["application/json"],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {
          "200": {"description": "Market details"},
          "404": {"description": "Market not found"}
        }
      }
    },
    "/api/v1/events": {
      "get": {
        "tags": ["Events"],
        "summary": "List all events",
        "produces": ["application/json"],
        "parameters": [
          {"name": "limit", "in": "query", "type": "integer", "default": 100},
          {"name": "cursor", "in": "query", "type": "string"},
          {"name": "active", "in": "query", "type": "boolean"}
        ],
        "responses": {
          "200": {"description": "List of events"}
        }
      }
    },
    "/api/v1/events/{id}": {
      "get": {
        "tags": ["Events"],
        "summary": "Get event by ID",
        "produces": ["application/json"],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {
          "200": {"description": "Event details"},
          "404": {"description": "Event not found"}
        }
      }
    },
...
    "/api/v1/positions": {
      "get": {
        "tags": ["User Data"],
        "summary": "Get user positions",
        "produces": ["application/json"],
        "parameters": [
          {"name": "address", "in": "query", "required": true, "type": "string"},
          {"name": "limit", "in": "query", "type": "integer", "default": 100}
        ],
        "responses": {
          "200": {"description": "List of positions"}
        }
      }
    },
    "/api/v1/top-movers": {
      "get": {
        "tags": ["Markets"],
        "summary": "Get top moving markets",
        "produces": ["application/json"],
        "parameters": [
          {"name": "limit", "in": "query", "type": "integer", "default": 10}
        ],
        "responses": {
          "200": {"description": "Top movers"}
        }
      }
    },
...
  },
...
}